Ajout d'un profil spring "local" pour éviter de dépendre de redis & elastic search
