+++ hesperides-spring/bootstrap/src/main/resources/application-local.yml
@@ -1,0 +1,3 @@
+#
+# Specific configuration for local profile.
+#
