+++ hesperides-spring/infrastructure/src/main/java/org/hesperides/infrastructure/auth/LocalWebSecurityConfig.java
@@ -1,0 +1,28 @@
+package org.hesperides.infrastructure.auth;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
+import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+
+@Configuration
+@Profile("local")
+public class LocalWebSecurityConfig extends WebSecurityConfigurerAdapter {
+
+    @Override
+    protected void configure(HttpSecurity http) throws Exception {
+        http
+                .authorizeRequests()
+                .anyRequest().fullyAuthenticated()
+                .and().httpBasic();
+    }
+
+    @Override
+    protected void configure(AuthenticationManagerBuilder auth) throws Exception {
+        auth.inMemoryAuthentication()
+                .withUser("tech").password("password").roles("TECH")
+                .and().withUser("prod").password("password").roles("PROD")
+                .and().withUser("user").password("password").roles("USER");
+    }
+}
