+++ hesperides-spring/infrastructure/src/main/java/org/hesperides/infrastructure/auth/WebSecurityConfig.java
@@ -22,6 +22,7 @@ package org.hesperides.infrastructure.auth;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
 import org.springframework.security.authentication.encoding.LdapShaPasswordEncoder;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
@@ -29,6 +30,7 @@ import org.springframework.security.config.annotation.web.configuration.EnableWe
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 
 @Configuration
+@Profile("!local")
 public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
     @Value("${ldap.urls}")
     private String ldapUrls;
