+++ hesperides-spring/infrastructure/src/main/java/org/hesperides/infrastructure/modules/elasticsearch/ModuleSearchRepository.java
@@ -18,7 +18,7 @@
  *
  *
  */
-package org.hesperides.infrastructure.elasticsearch;
+package org.hesperides.infrastructure.modules.elasticsearch;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
